requirements.txt
# Core ML
scikit-learn==1.4.0
numpy==1.26.4
pandas==2.2.0

# Deep Learning
torch==2.2.0
transformers==4.38.0
datasets==2.17.0
accelerate==0.27.0

# NLP / Text processing
nltk==3.8.1
spacy==3.7.2
beautifulsoup4==4.12.3
langdetect==1.0.9

# Mastodon API
Mastodon.py==1.8.1
requests==2.31.0

# Evaluatie en visualisatie
matplotlib==3.8.2
seaborn==0.13.2
scipy==1.12.0
statsmodels==0.14.1
plotly==5.18.0

# Utilities
tqdm==4.66.1
python-dotenv==1.0.1
joblib==1.3.2
pyyaml==6.0.1
loguru==0.7.2

# Development / Testing
pytest==8.0.0
pytest-cov==4.1.0
jupyter==1.0.0
ipykernel==6.29.0
black==24.1.1
flake8==7.0.0
